Move SubError to BaseException. Refactoring some code
